util.rs: move some code to function to reuse
- Move code to func `parse_rpm_metadata` to parse the output of `rpm -q`
- Move code to func `rpm_query` to query rpm database, list the
package and build times
